+++ lib/src/elements/command_options/th_point_height_value_command_option.dart
@@ -167,7 +167,9 @@ class THPointHeightValueCommandOption extends THCommandOption
       asString += '?';
     }
 
-    if (unitSet) {
+    if (unitSet &&
+        (mode != THPointHeightValueMode.presumedMinus) &&
+        (mode != THPointHeightValueMode.presumedPlus)) {
       asString += " $unit";
     }